+++ mayan/apps/ocr/backends/literals.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, unicode_literals
+
+DEFAULT_TESSERACT_BINARY_PATH = '/usr/bin/tesseract'
+DEFAULT_TESSERACT_TIMEOUT = 600  # 600 seconds, 10 minutes